requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.3.1
APScheduler==3.6.3
async-timeout==3.0.1
attrs==22.2.0
Brotli==1.0.9
cachetools==4.2.2
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==2.1.1
colorama==0.4.6
discord==2.1.0
discord.py-self==1.9.2
filetype==1.2.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
python-dotenv==0.21.0
python-telegram-bot==13.15
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-toolbelt==0.10.1
six==1.16.0
tornado==6.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
ua-parser==0.16.1
urllib3==1.26.13
websocket-client==0.59.0
yarl==1.8.2
